Project completed

+++ server/app/controllers/todo_controller.py
@@ -1,11 +1,12 @@
 import app.database as database
 
 # THIS IS GET TODO CONTROLLER
-async def get_all_todo():
+async def get_all_todo(current_user):
     async with database.pool.acquire() as conn:
         async with conn.cursor() as cursor:
             await cursor.execute(
-                "SELECT * FROM todos"
+                "SELECT * FROM todos WHERE user_id=%s",
+                (current_user.get("id"),)
             )
 
             todos = await cursor.fetchall()   
@@ -14,20 +15,23 @@ async def get_all_todo():
 
 
 # THIS IS ADD TODO CONTROLLER
-async def add_todo(todo):
+async def add_todo(todo,current_user):
     async with database.pool.acquire() as conn:
         async with conn.cursor() as cursor:
-           await cursor.execute(
-               "INSERT INTO todos (title) VALUES (%s)",(todo.title,)
-           )
-    
+            try:
+                await cursor.execute(
+               "INSERT INTO todos (title,user_id) VALUES (%s,%s)",(todo.title,current_user.get("id")))
+            except Exception as e:
+                print(str(e))
+                
     return {
         "message": "Todo added succesfully"
     }
 
 
 # THIS IS UPDATE TODO CONTROLLER
-async def update_todo(id,todo):
+async def update_todo(id,todo,current_user):
+    
     update_data = todo.model_dump(exclude_unset=True)
     print(update_data)
     
@@ -42,8 +46,9 @@ async def update_todo(id,todo):
         values.append(val)
 
     values.append(id)
+    values.append(current_user.get("id"))
 
-    query = f"UPDATE todos SET {', '.join(fields)} WHERE id=%s"
+    query = f"UPDATE todos SET {', '.join(fields)} WHERE id=%s AND user_id=%s"
     
     async with database.pool.acquire() as conn:
         async with conn.cursor() as cursor:
@@ -55,11 +60,11 @@ async def update_todo(id,todo):
 
 
 # THIS IS DELETE TODO CONTROLLER
-async def delete_todo(id):
+async def delete_todo(id,current_user):
     async with database.pool.acquire() as conn:
         async with conn.cursor() as cursor: 
             await cursor.execute(
-                "DELETE FROM todos WHERE id=%s", (id,)
+                "DELETE FROM todos WHERE id=%s AND user_id=%s", (id,current_user.get("id"))
             )
     
     return {
